fix(engine): resolve all clippy warnings (clamp patterns, unused imports)
Replace manual clamp patterns with .clamp() calls across cerebellar,
channels, and population modules. Fix unused import (spike_times),
unnecessary casts, manual swap patterns, and other clippy warnings
in analysis and neuron modules.

cargo clippy -- -D warnings now passes cleanly.

+++ engine/src/neurons/hardware.rs
@@ -380,7 +380,7 @@ impl AkidaNeuron {
         if self.spiked {
             return 0;
         }
-        self.v += (weight as f64 * self.modulation.powi(self.rank)) as i32;
+        self.v += (weight * self.modulation.powi(self.rank)) as i32;
         self.rank += 1;
         if self.v >= self.threshold {
             self.spiked = true;
